Update: Change orderBy DESC createdAt + return Page in RoomTypeServiceImpl

+++ src/main/java/vn/eledevo/vksbe/service/roomtype/RoomTypeServiceImpl.java
@@ -3,8 +3,8 @@ package vn.eledevo.vksbe.service.roomtype;
 import static vn.eledevo.vksbe.constant.ResponseMessage.ROOM_TYPE_EXISTED;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -19,6 +19,7 @@ import vn.eledevo.vksbe.entity.RoomType;
 import vn.eledevo.vksbe.entity.User;
 import vn.eledevo.vksbe.exception.ValidationException;
 import vn.eledevo.vksbe.mapper.RoomTypeMapper;
+import vn.eledevo.vksbe.repository.RoomRepository;
 import vn.eledevo.vksbe.repository.RoomTypeRepository;
 import vn.eledevo.vksbe.utils.SecurityUtils;
 
@@ -29,11 +30,13 @@ public class RoomTypeServiceImpl implements RoomTypeService {
 
     RoomTypeRepository repository;
 
+    RoomRepository roomRepository;
+
     RoomTypeMapper mapper;
 
     @Override
     public List<RoomTypeResponse> getAllRoomType() {
-        List<RoomType> roomTypeListFromDB = repository.findAll();
+        List<RoomType> roomTypeListFromDB = repository.findAll(Sort.by(Sort.Direction.DESC, "createdAt", "updatedAt"));
         List<RoomTypeResponse> roomTypeList =
                 roomTypeListFromDB.stream().map(mapper::toResponse).toList();
         return roomTypeList;
@@ -49,11 +52,11 @@ public class RoomTypeServiceImpl implements RoomTypeService {
         user.setId(SecurityUtils.getUserId());
 
         if (repository.existsByName(roomTypeRequest.getName())) {
-            throw new ValidationException("Tên phòng", ROOM_TYPE_EXISTED);
+            throw new ValidationException("nameExist", ROOM_TYPE_EXISTED);
         }
 
         roomType.setName(roomTypeRequest.getName());
-        roomType.setMaxPeople(roomTypeRequest.getMaxPeople());
+        roomType.setMaxPeople(String.valueOf(roomTypeRequest.getMaxPeople()));
         roomType.setDescription(roomTypeRequest.getDescription());
         roomType.setUser(user);
 
@@ -77,7 +80,7 @@ public class RoomTypeServiceImpl implements RoomTypeService {
         }
 
         roomType.setName(roomTypeRequest.getName());
-        roomType.setMaxPeople(roomTypeRequest.getMaxPeople());
+        roomType.setMaxPeople(String.valueOf(roomTypeRequest.getMaxPeople()));
         roomType.setDescription(roomTypeRequest.getDescription());
         roomType.setUser(user);
 
@@ -90,25 +93,24 @@ public class RoomTypeServiceImpl implements RoomTypeService {
     public RoomTypeResponse deleteRoomType(Long id) throws ValidationException {
         RoomType roomType =
                 repository.findById(id).orElseThrow(() -> new ValidationException("Room type", "Room type not found!"));
+
+        if (roomRepository.existsByRoomTypeId(id)) {
+            throw new ValidationException("errorDelete", "Loại phòng này đang được sử dụng! Không thể xóa!");
+        }
+
         repository.deleteById(id);
         return null;
     }
 
     @Override
-    public List<RoomTypeResponse> filterRoomType(
+    public Page<RoomTypeResponse> filterRoomType(
             String orderBy, int page, int limit, String orderedColumn, String name, String maxPeople)
             throws ValidationException {
         Pageable roomTypePageable =
                 PageRequest.of(page - 1, limit, Sort.by(Sort.Direction.valueOf(orderBy.toUpperCase()), orderedColumn));
 
-        List<RoomType> roomTypeList = repository.listRoomTypeSearchedAndPagingFromDB(name, maxPeople, roomTypePageable);
+        Page<RoomType> roomTypeList = repository.listRoomTypeSearchedAndPagingFromDB(name, maxPeople, roomTypePageable);
 
-        if (roomTypeList.isEmpty()) {
-            throw new ValidationException("Trống", "Không tìm thấy kiểu phòng tương ứng!");
-        }
-
-        List<RoomTypeResponse> roomTypeResponseList =
-                roomTypeList.stream().map(mapper::toResponse).collect(Collectors.toList());
-        return roomTypeResponseList;
+        return roomTypeList.map(mapper::toResponse);
     }
 }
